+++ grand_trade_auto/orm/orm_meta.py
@@ -62,11 +62,13 @@ class Orm(ABC):
         """
         self._create_schema_enum_currency()
         self._create_schema_enum_market()
+        self._create_schema_enum_price_frequency()
 
         self._create_schema_table_datafeed_src()
         self._create_schema_table_exchange()
         self._create_schema_table_company()
         self._create_schema_table_security()
+        self._create_schema_table_security_price()
         self._create_schema_table_stock_adjustment()
 
 
@@ -90,6 +92,20 @@ class Orm(ABC):
         """
         Create the market enum.  The values must all match exactly the values as
         shown in model_meta.
+
+        Dependent on: None
+        Dependent on tables: N/A
+
+        Subclass must define and execute SQL/etc.
+        """
+
+
+
+    @abstractmethod
+    def _create_schema_enum_price_frequency(self):
+        """
+        Create the price_frequency enum.  The values must all match exactly the
+        values as shown in model_meta.
 
         Dependent on: None
         Dependent on tables: N/A
@@ -145,6 +161,19 @@ class Orm(ABC):
 
         Dependent on enums: currency, market
         Dependent on tables: company, datafeed_src, exchange
+
+        Subclass must define and execute SQL/etc.
+        """
+
+
+
+    @abstractmethod
+    def _create_schema_table_security_price(self):
+        """
+        Create the security_price table.
+
+        Dependent on enums: price_frequency
+        Dependent on tables: datefeed_src, security
 
         Subclass must define and execute SQL/etc.
         """