+++ tests/unit/orm/test_postgres_orm.py
@@ -130,6 +130,7 @@ def _test_create_schema_enum(orm, test_func, enum_name,
 @pytest.mark.parametrize('method_name, enum_name', [
     ('_create_schema_enum_market', 'market'),
     ('_create_schema_enum_currency', 'currency'),
+    ('_create_schema_enum_price_frequency', 'price_frequency'),
 ])
 def test__create_schemas_enums(pg_test_orm, method_name, enum_name):
     """
@@ -212,6 +213,7 @@ def _test_create_schema_table(orm, test_func, table_name,
     ('_create_schema_table_exchange', 'exchange'),
     ('_create_schema_table_company', 'company'),
     ('_create_schema_table_security', 'security'),
+    ('_create_schema_table_security_price', 'security_price'),
     ('_create_schema_table_stock_adjustment', 'stock_adjustment'),
 ])
 def test__create_schemas_tables(pg_test_orm, method_name, table_name):